Use C++11 universal brace-initialization syntax everywhere.
Given that we've just changed
    Entity e;
    ZERO(&e);
to
    Entity e = {};
this may seem like pointless churn. This is, however, not the case.

The change above deals with the fact that it is not legal to use
zero out the underlying representation to initialize the non-POD
types; in absence of any nontrivial constructors, this change is
a faithful translation of the old intent behind that code while taking
into account that members now may have non-POD types.

In this commit, we are changing
    Entity e = {};
to
    Entity e {};
which may (unfortunately) seem more trivial than it is.

The reason we are making this change is that we would like to use
nontrivial constructors in some of our types. Ideally we could write
the variable declaration of type Entity with zero-initialization as
    Entity e();
but this is a well-known pitfall in the C++ syntax: the code above
does not declare a variable of type Entity called "e" but rather
it declares a function called "e" that accepts no arguments and
returns an Entity; this is why we bothered with Entity e = {};

Now, we would like to introduce a constructor with an argument--
perhaps something like
    class Entity { Entity(Sketch *sketch); }
We could then write the declaration of a variable of type Entity
that is initialized using the constructor above as
    Entity e = {sketch};
using the same kind of syntax that we used for zero-initialization
before. However, there are several issues:

The first issue is that this results in an extraneous copy.
First, an Entity(Sketch*) constuctor is called to construct
a temporary. Second, an Entity(const Entity&) copy construtor
is called to construct the variable. Many compilers will
optimize this out via copy elision, however...

The second issue is that this interacts badly with explicit
constructors. As we've seen, a temporary is first constructed,
and this is exactly the behavior that is prohibited by declaring
the constructor explicit. Thus this syntax is fragile.

By writing this as
    Entity e {sketch};
we avoid all of the issues described above, since it calls the right
constructor directly.

There are more benefits to this syntax in general, as it is rightly
called "universal": it can be used in any context (variable
declaration; constructor initializer list; default member initializer;
new operator expressions; etc) and it can be used equally well with
any types: POD types, types with user-defined constructors, even when
initializing collections as
    std::vector<int> v = {1, 2, 3};

Since we already require C++11, and the universal brace-initialization
syntax has clear benefits, we switch all initializations to use it.